Replaced LiftThurst Calculation with Direct Thruster output.
I replaced the thruster output to a more direct power output value of all downwards facing thrusters. in particular which thruster is struggling the most to keep the ship afloat. this gives a much more accurate measure of if the ship can stay afloat or not and it will now reach 100% if you run out of hover power in any direction so it will account for drifting at odd angles too. all in all it gives a much more reliable and more understandable measure of how much power you have left at your disposal.